requirements.txt
affine==2.4.0
aiobotocore==2.15.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
apidaora==0.28.0
asttokens==2.4.1
async-property==0.2.2
async-timeout==4.0.3
asyncapi==0.14.1
asyncio-redis==0.16.0
attrs==24.2.0
botocore==1.35.36
Bottleneck==1.4.2
bounded-pool-executor==0.0.3
broadcaster==0.3.1
cached-property==1.5.2
certifi==2024.7.4
cffi==1.17.1
cftime==1.6.4.post1
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
dask==2024.11.1
debugpy==1.8.2
decorator==5.1.1
deprecation==2.1.0
dictdaora==0.0.2
earthaccess==0.11.0
ecdsa==0.19.0
exceptiongroup==1.2.2
executing==2.0.1
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.10.0
GDAL==3.6.2
geopandas==1.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.7
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jplephem==2.22
jsondaora==0.14.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jwcrypto==1.5.6
kiwisolver==1.4.5
locket==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
multidict==6.1.0
multimethod==1.12
nest-asyncio==1.6.0
netCDF4==1.6.2
nost_tools==1.1.5
ntplib==0.4.0
numexpr==2.10.1
numpy==1.26.4
orjson==3.10.7
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pika==1.3.2
pillow==10.4.0
pip==24.3.1
platformdirs==4.2.2
plotly==5.24.0
pqdm==0.2.0
prompt_toolkit==3.0.47
propcache==0.2.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pycparser==2.22
pydantic==2.9.1
pydantic-asyncapi==0.2.1
pydantic_core==2.23.3
Pygments==2.18.0
pyhdf==0.11.4
pyogrio==0.10.0
pyotp==2.9.0
pyparsing==3.2.0
PyPDF2==3.0.1
pyproj==3.7.0
python-cmr==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==4.2.3
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
rasterio==1.4.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
rioxarray==0.17.0
rsa==4.9
s3fs==2024.10.0
scipy==1.14.1
setuptools==74.1.2
sgp4==2.23
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
skyfield==1.49
sniffio==1.3.1
snuggs==1.4.7
stack-data==0.6.3
tenacity==9.0.0
tinynetrc==1.3.1
toolz==1.0.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
wheel==0.45.0
wordcloud==1.9.3
wrapt==1.16.0
xarray==2023.6.0
yarl==1.17.1
zipp==3.21.0
